requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
fake-useragent==1.2.1
h11==0.14.0
idna==3.4
importlib-resources==6.0.1
outcome==1.2.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.4
undetected-chromedriver==3.5.3
urllib3==2.0.4
websockets==11.0.3
wsproto==1.2.0
zipp==3.17.0
